fix: implemented opacity and z-index transition feedback
