+++ packages/epo-react-lib/src/atomic/SlideoutInfoCard/index.tsx
@@ -1,25 +1,2 @@
-import {ReactNode} from 'react';
-import * as Styled from "./styles";
+export {default} from "./SlideoutInfoCard";
 
-type SlideFrom = "top" | "right" | "bottom" | "left";
-interface SlideoutInfoCardProps {
-  isOpen?: boolean;
-  slideFrom?: SlideFrom;
-  className?: string;
-  children: ReactNode,
-};
-
-function SlideoutInfoCard({
-  isOpen = false, 
-  slideFrom = "left", 
-  children
-} : SlideoutInfoCardProps) {
-
-  return (
-    <Styled.SlideoutInfoCard isOpen={isOpen} slideFrom={slideFrom}>
-      {children}
-    </Styled.SlideoutInfoCard>
-  );
-}
-
-export default SlideoutInfoCard;
