+++ packages/epo-widget-lib/src/widgets/OrbitalSim/OrbitalSim.tsx
@@ -85,23 +85,22 @@ function OrbitalSim() {
   let isDisabled = false;
   return (
     <>
-       <Styled.GlobalStyles/>
-       <Styled.OrbitalSimWrapper>
-        {swappableOrbits ? (
-          <Styled.SwappableOrbitsContainer>
-            {orbits && orbits.neos && orbits.neos.map((neo: Neo) => (
-              <Styled.SwappableOrbitButton 
-                onClick={() => updateSwappableOrbit(neo)} 
-                data-active={
-                  ((Object.keys(currentSwappableOrbit).length) ? (neo.Principal_desig === currentSwappableOrbit.neos[0].Principal_desig) : false)
-                }>
-                  {neo.Ref}
-              </Styled.SwappableOrbitButton>
-            ))}
-          </Styled.SwappableOrbitsContainer>
-          
-        ) : (
-          <>
+      <Styled.GlobalStyles/>
+      <Styled.OrbitalSimWrapper>
+      {swappableOrbits ? (
+        <Styled.SwappableOrbitsContainer>
+          {orbits && orbits.neos && orbits.neos.map((neo: Neo) => (
+            <Styled.SwappableOrbitButton 
+              onClick={() => updateSwappableOrbit(neo)} 
+              data-active={
+                ((Object.keys(currentSwappableOrbit).length) ? (neo.Principal_desig === currentSwappableOrbit.neos[0].Principal_desig) : false)
+              }>
+                {neo.Ref}
+            </Styled.SwappableOrbitButton>
+          ))}
+        </Styled.SwappableOrbitsContainer>
+      ) : (
+        <>
         { detailsRows && showDetailsTable && (
           <OrbitalDetails />
         )}
@@ -126,7 +125,7 @@ function OrbitalSim() {
         </>
       )}
        
-       <Styled.CanvasWrapper orthographic={true}>
+      <Styled.CanvasWrapper orthographic={true}>
             <CameraController {...{ pov: allowOrbitRotation ? null : ( pov ?? "top"), reset }} />
             <Camera
               near={-1000}
